fix(cron): the check read the invoking user's crontab, not the owner's (Refs #348)
A cron resource with no `owner:` installs into ROOT's crontab. Its check read
the SSH user's. The resource was correctly installed and permanently
unconvergeable, and every dependent was skipped.

Measured on paiml's intel:

    $ crontab -l | grep -c ci-image-rebuild
    0                                   <- what the check looked at
    $ sudo crontab -l | grep rebuild.sh
    30 3 * * * bash .../rebuild.sh      <- where the apply had put it

    JIDOKA: intel/ci-image-rebuild failed - dependents will be skipped:
      apply exited 0 but the host does not report the declared state (check exit 1)

CAUSE. `check_script`, `apply_script` and `state_query_script` each re-derived
their crontab command independently, and only the apply carried the
`SUDO=""` / `[ "$(id -u)" -ne 0 ] && SUDO="sudo"` preamble. They agreed on the
owner (`root`); they disagreed on the PRIVILEGE.

Reading another user's crontab is exactly as privileged as writing it.
`crontab -u <user>` refuses EVERY non-root caller — cronie exits with "must be
privileged to use -u" even for the caller's own username. `2>/dev/null`
swallowed that refusal, `grep -qF` then read an empty stream and exited 1, and
`verdict::single` turned that into the divergence `output_verify` reports.

So the check could not distinguish "the job is not installed" from "I was not
allowed to look", and asserted the first.

The same omission in `state_query_script` is the more expensive half: the
observable recorded `cron=MISSING:<name>` for a job that exists, so the lock
stored "absent" as the OBSERVED state and drift was wrong in the same
direction.

Fixed by making one function decide and having all three call sites delegate:
`crontab_user()` for the identity, `SUDO_PREAMBLE` for the privilege. The
apply's emitted bytes are unchanged, so no apply behaviour and no apply-side
script hash moves.

`CRONTAB_CHECK_GUARD` is the second half and is not optional. `crontab -l`
exits 1 for BOTH "no crontab for user" and EPERM, so on a host with no
passwordless sudo the false `missing:` would simply move one step later. The
guard takes the honest signal BEFORE the read and exits 2, which `cli::check`
maps to SKIP and `output_verify` treats as neither converged nor diverged.
Modelled on `service::SYSTEMD_CHECK_GUARD` and `network::UFW_CHECK_GUARD`.

REPORTING CHANGE, and the correct one: a host without passwordless sudo moves
from a false `missing:`/FAIL to SKIP. Anyone counting cron resources as failing
will see that count move. This cannot hang — `stdin_isolation` gives the whole
script `< /dev/null`, ssh uses `BatchMode=yes` with no `-t`, and `sudo -n` never
prompts.

The docs already promised this and the code never did: 03-resources.md's lint
table said "`$SUDO` in crontab read/write" while the read half did not exist,
and the state-query example showed a bare `crontab -l -u root`. Both corrected.

cron.rs was EXACTLY 500 lines, so the inline `mod tests` moves verbatim to
src/resources/tests_cron_b.rs, leaving cron.rs at 110. The moved assertions are
strengthened from `crontab -u '<user>' -l` to `$SUDO crontab -u '<user>' -l`;
the old substring survives the prefix, which is precisely why no existing test
caught this.

tests/falsification_cron_check_reads_the_owner_crontab.rs EXECUTES the generated
scripts against a fake `id`/`sudo`/`crontab` reproducing that host: a non-root
caller, a working sudo, and a `crontab` that refuses `-u` to anyone who did not
come through it. All four tests fail on the pre-fix code — the check exits 1
with `missing:ci-image-rebuild`, the state query prints
`cron=MISSING:ci-image-rebuild` for a job that is right there, and the no-sudo
host exits 1 instead of 2.
