+++ src/resources/mod.rs
@@ -38,6 +38,8 @@ mod tests_build;
 #[cfg(test)]
 mod tests_cron;
 #[cfg(test)]
+mod tests_cron_b;
+#[cfg(test)]
 mod tests_docker;
 #[cfg(test)]
 mod tests_docker_b;
